% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{barplot_ranef}
\alias{barplot_ranef}
\title{Plot random effect estimates for host species}
\usage{
barplot_ranef(
  model,
  host_spp = NULL,
  interaction_df = NULL,
  species_col = NULL,
  host_col,
  host_status_col = NULL,
  species,
  colour = "coral",
  cex_names = 0.5,
  ylab = "Random effect estimates",
  main = species
)
}
\arguments{
\item{model}{A fitted Bayesian model of class \code{brmsfit}.}

\item{host_spp}{A character string specifying the host species. This can be
provided as an alternative to inferring the host species via `get_host_spp`
when an `interaction_df` is supplied.  Default: `NULL` (not required if
`interaction_df` is provided).}

\item{interaction_df}{A data frame containing species-host interaction data.
Default: `NULL` (not required if `host_spp` is provided).}

\item{species_col}{A character string indicating the column name in
`interaction_df` that contains species names.  Default: `NULL` (not required
if `interaction_df` is provided).}

\item{host_col}{A character string indicating the column name in both `model`
and `interaction_df` that contains host species names.}

\item{host_status_col}{A character string specifying the column in
`interaction_df` that indicates the host status.  Default: `NULL` (not
required if `interaction_df` is provided).}

\item{species}{Character string giving the focal species of interest.}

\item{colour}{Character string specifying the bar colour. Default: `coral`.}

\item{cex_names}{Numeric value controlling the size of axis labels. Default:
`0.5`.}

\item{ylab}{Character string specifying the y-axis label. Default:
`"Random effect estimates"`.}

\item{main}{Character string indicating plot title. Defaults to the name
of the focal species (set to `NULL` to ignore).}
}
\value{
Invisibly returns the random effect estimates used in the barplot.
}
\description{
This function extracts and plots random effect estimates from a
\code{brmsfit} model for host species associated with a given focal species.
}
\examples{
plot_ranef(interaction_df = interaction_data,
           model = oak_model,
           species = "Agrilus coxalis",
           species_col = "agrilus_sp",
           host_col = "quercus_sp",
           host_status_col = "interaction")

}
\seealso{
\code{\link{get_host_spp}}, \code{brms::ranef}
}
