% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.r
\name{generate_formulas}
\alias{generate_formulas}
\title{Generate all possible model formulas from a set of variables}
\usage{
generate_formulas(variables, response, include_null = TRUE)
}
\arguments{
\item{variables}{A character vector of predictor variables (right-hand
side terms).}

\item{response}{A single character string specifying the response variable
(left-hand side term).}

\item{include_null}{Logical; if `TRUE`, a null model (\code{response ~ 1})
is included as the first formula.}
}
\value{
A list of \code{formula} objects representing all model combinations.
}
\description{
This function creates all possible combinations of predictor variables
for use in model formulae, optionally including a null (intercept-only)
model. The resulting list of formula objects can be used directly in
modelling functions.
}
\examples{
\dontrun{
vars <- c("geo.dist",
          "phylo.dist",
          "(1 | gr(plant_sp, cov = phylo_cov))")

models <- generate_formulas(
  variables = vars,
  response = "interaction",
  include_null = TRUE
)
}

}
