% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{get_gbif_keys}
\alias{get_gbif_keys}
\title{Retrieve GBIF taxon keys for a list of species}
\usage{
get_gbif_keys(species, higher_taxon = "", rank = "SPECIES", ...)
}
\arguments{
\item{species}{Character vector of species names to look up.}

\item{higher_taxon}{Character string specifying a higher taxon key to
restrict the search.}

\item{rank}{Character string specifying the taxonomic rank to query
(default "SPECIES").}

\item{...}{Additional arguments passed directly to `rgbif::name_lookup()`.}
}
\value{
A data.frame with two columns:
  \describe{
    \item{species}{Original species name.}
    \item{key}{Corresponding GBIF taxon key (NA if not found).}
  }
}
\description{
This function is a lightweight wrapper around `rgbif::name_lookup()`. It
takes a character vector of species names and queries GBIF to retrieve their
GBIF taxon keys, and returns a data frame indicating the GBIF key for each
species. Some species may not be found and will require manual lookup.
Function allows manual overriding of any parameters in
`name_lookup()` via `...`.
}
\examples{
# Simple example using dummy species
\dontrun{
species_list <- c("Quercus robur", "Quercus alba")
get_gbif_keys(species_list, higher_taxon = "220")
# Override `rgbif::name_lookup` `status` parameter
get_gbif_keys(species_list, higher_taxon = "220", status = "Synonym")
}

}
