% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{get_host_spp}
\alias{get_host_spp}
\title{Extract host species for a given focal species}
\usage{
get_host_spp(df, species_col, host_col, host_status_col, species)
}
\arguments{
\item{df}{A data frame containing species–host interaction data.}

\item{species_col}{Character string specifying the column containing species
names.}

\item{host_col}{Character string specifying the column containing host
species names.}

\item{host_status_col}{Character string specifying the column indicating
host status (values must include 0 and 1, where 1 indicates a host).}

\item{species}{Character string giving the focal species of interest.}

\item{main}{Character string indicating plot title. Defaults to the name
of the focal species (set to `NULL` to ignore).}
}
\value{
A character vector of unique host species associated with the focal
species.
}
\description{
This function identifies host species associated with a focal species
from a species–host interaction data frame.
}
\keyword{internal}
