% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{plot_fourfold}
\alias{plot_fourfold}
\title{Plot a fourfold confusion matrix and optionally print performance metrics}
\usage{
plot_fourfold(observations, predictions, print_metrics = TRUE)
}
\arguments{
\item{print_metrics}{Logical; if \code{TRUE} (default), sensitivity and
specificity are calculated and printed.}

\item{observed_status}{A factor or numeric vector containing observations,
coded as 0 and 1.}

\item{prediction_binary}{A factor or numeric vector containing predictions
, coded as 0 and 1.}
}
\value{
Generates a fourfold plot and, optionally, prints classification
metrics.
}
\description{
Generates a fourfold plot of a binary confusion matrix using predicted and
observed class labels. The function also computes and prints sensitivity
(true positive rate) and specificity (true negative rate) if requested.
Input vectors may be factors or numeric, containing only 0s and 1s.
}
\examples{
\dontrun{
obs <- c(0,1,1,0,1)
pred <- c(0,1,0,0,1)
plot_fourfold(obs, pred)
}

}
