% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{rename_labels}
\alias{rename_labels}
\title{Rename tip labels in a phylogeny}
\usage{
rename_labels(tree, old_labels, new_labels)
}
\arguments{
\item{tree}{Phylogeny of class \code{phylo} containing \code{tip.label}.}

\item{old_labels}{Character vector of tip labels to be replaced.}

\item{new_labels}{Character vector of new tip labels to replace the old ones.}
}
\value{
A \code{phylo} tree with specified tip labels renamed.
}
\description{
Replaces specified tip labels in a \code{phylo} tree with new labels.
Both \code{old_labels} and \code{new_labels} must be character vectors
of the same length.
}
\keyword{internal}
