% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{standardise_phylo}
\alias{standardise_phylo}
\title{Standardise the tip labels of a phylogeny}
\usage{
standardise_phylo(
  tree,
  clean_labels = TRUE,
  pattern,
  replacement = "\\\\1. \\\\2",
  remove_duplicates = TRUE,
  old_labels = NULL,
  new_labels = NULL
)
}
\arguments{
\item{tree}{A phylogeny of class \code{phylo}.}

\item{clean_labels}{Logical; if \code{TRUE}, clean tip labels (for plotting purposes).}

\item{pattern}{Character; regex pattern for extracting species-level labels.}

\item{replacement}{Character; replacement pattern for use in \code{gsub}.}

\item{remove_duplicates}{Logical; if \code{TRUE}, remove duplicated tip labels.}

\item{old_labels}{Character vector of labels to rename (passed to internal \code{rename_labels}).}

\item{new_labels}{Character vector of new labels to use (passed to internal \code{rename_labels}).}
}
\value{
A \code{phylo} tree with standardised tip labels.
}
\description{
Applies standardisation steps to a \code{phylo} tree's tip labels, including:
\itemize{
  \item Cleaning tip label names (using regex).
  \item Renaming specified tip labels.
  \item Removing duplicate tip labels, keeping only one instance.
}
It returns a phylogeny.
}
\examples{
\dontrun{
tree <- ape::read.tree("path/to/tree.tre")
tree <- standardise_phylo(tree,
                          clean_labels = TRUE,
                          pattern = "^([A-Z])[a-z]+_*([a-z]+).*",
                          replacement = "\\\\1. \\\\2",
                          remove_duplicates = TRUE,
                          old_labels = "Quercus maxima",
                          new_labels = "Quercus rubra")
plot(tree)
}

}
