% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.r
\name{tpr}
\alias{tpr}
\title{Compute the true positive rate (TPR)}
\usage{
tpr(predictions, observations = rep(1, length(predictions)))
}
\arguments{
\item{predictions}{Numeric vector of predicted classes, expected to contain
only \code{0} and \code{1}.}

\item{observations}{Numeric vector of observed classes, expected to contain
only \code{0} and \code{1}. If not supplied, all observations default to
\code{1}. Defaults to \code{rep(1, length(predictions))}.}
}
\value{
A numeric value giving the true positive rate as a percentage,
rounded to two decimal places.
}
\description{
Calculates the true positive rate from a vector of binary predictions and a
corresponding vector of binary observations. The true positive rate is
defined as the proportion of observed positives correctly predicted.
}
\details{
The function assumes that both input vectors contain only binary values,  abd
calculates the true positive rate as:

\deqn{TPR = \frac{TP}{TP + FN} \times 100}

The function assumes that both input vectors contain only binary values.
}
\examples{
\dontrun{
tpr(c(1, 0, 1, 1), c(1, 1, 1, 0))
}

}
