% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{barplot_host_distance}
\alias{barplot_host_distance}
\title{Plot phylogenetic distances for host species}
\usage{
barplot_host_distance(
  interaction_df,
  species,
  species_col,
  host_col,
  host_status_col,
  dist_col,
  host_spp = NULL,
  colour = "coral",
  cex_names = 0.5,
  ylab = "Distance to another host species",
  main = species
)
}
\arguments{
\item{interaction_df}{A data frame containing species–host interaction data.}

\item{species}{Character string giving the focal species of interest.}

\item{species_col}{Character string specifying the column containing species
names in `interaction_df`.}

\item{host_col}{Character string specifying the column containing host
species names in `interaction_df`.}

\item{host_status_col}{Character string specifying the column indicating host
status in `interaction_df`.}

\item{dist_col}{Character string specifying the column containing
phylogenetic distances in`interaction_df`.}

\item{host_spp}{A character string specifying the host species. This can be
provided as an alternative to inferring the host species via `get_host_spp`
when an `interaction_df` is supplied.  Default: `NULL` (not required if
`interaction_df` is provided).}

\item{colour}{Character string specifying the bar colour. Default: `"coral"`.}

\item{cex_names}{Numeric value controlling the size of axis labels. Default:
`0.5`.}

\item{ylab}{Character string specifying the y-axis label. Default:
`"Distance to another host species"`.}

\item{main}{Character string indicating plot title. Defaults to the name
of the focal species (set to `NULL` to ignore).}
}
\value{
Invisibly returns the heights used in the barplot.
}
\description{
This function creates a barplot showing phylogenetic distances
associated with host species for a given focal species.
}
\examples{
\dontrun{
barplot_host_distance(interaction_df = interaction_data,
                      species = "Agrilus coxalis",
                      species_col = "agrilus_sp",
                      host_col = "quercus_sp",
                      host_status_col = "interaction",
                      dist_col = "phylo_dist_min")
}

}
\seealso{
\code{\link{get_host_spp}}
}
