% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.r
\name{generate_geo_metrics_df}
\alias{generate_geo_metrics_df}
\title{Calculate mean and median minimum geographic distance metrics}
\usage{
generate_geo_metrics_df(
  expanded_interaction_df,
  coords_df,
  known_interactions_df,
  host_taxon_col,
  hosted_taxon_col,
  subsample_coords = 10000,
  measure = "cheap",
  metric_type = c("mean", "median", "norm", "log"),
  dnorm_args = c(mean = 0, sd = 50),
  dnorm_from = c(min = 0, max = 1e+05),
  round_val = 2,
  max_dist_km = 30000,
  verbose = TRUE,
  geodist_quiet = TRUE
)
}
\arguments{
\item{expanded_interaction_df}{Data frame of all possible host-taxon-member
– hosted species interactions to evaluate. Must include columns for host and
hosted taxa.}

\item{coords_df}{Data frame of location occurrences for all host-taxon-member
species. Column names must be "lon", "lat". Must include columns for host
taxa, longitude (`lon`), and latitude (`lat`).}

\item{known_interactions_df}{Data frame of known "positive" interactions.
Must include columns for host and hosted taxa.}

\item{host_taxon_col}{Character. Column name indicating the host taxon in all
input data frames.}

\item{hosted_taxon_col}{Character. Column name indicating the hosted taxon in
`expanded_interaction_df` and `known_interactions_df`.}

\item{subsample_coords}{Integer or NULL. Number of geographic records to
sample per host taxon to reduce computational load. Defaults to `10000`. Set
to NULL to use all coordinates.}

\item{measure}{Character. Distance calculation method passed to
`geodist::geodist`. Options: `"cheap"`, `"haversine"`, `"vincenty"`, or
`"geodesic"`. Defaults to `"cheap"`.}

\item{metric_type}{Character vector specifying which distance summaries and
transformations to compute. Options: `"mean"`, `"median"`, `"norm"`, `"log"`.
`"mean"` and/or `"median"` control which base metrics are calculated, while
`"norm"` and `"log"` apply normalisation and log-transformation (via
[transform_distances()]) to those metrics. Defaults to
`c("mean", "median", "norm", "log")`. At least one of `"mean"` or `"median"`
must be included.}

\item{dnorm_args}{Named numeric vector giving the `mean` and `sd` parameters
used by transform_distances()] for normalisation. Defaults to
`c(mean = 0, sd = 50)`.}

\item{dnorm_from}{Numeric vector of two values giving the range (in km) used
by [transform_distances()] to anchor the scaling of the normalised scores.
Defaults to `c(0, 1e5)`.}

\item{round_val}{Numeric. Integer indicating the number of decimal places
used for rounding results. Detaults to `2`.}

\item{max_dist_km}{Numeric. Maximum distance (in km) to assign when no other
hosts exist for a hosted species other than the curent taxon. Defaults to
`30000`.}

\item{verbose}{Logical. Set to false to ignore warnings and messages.
Defaults to `TRUE`.}

\item{quiet}{Logical. Set to true to ignore `geodist::geodist` (recommended
if `measure = "cheap"`.). Defaults to `TRUE`.}
}
\value{
A data frame identical to `expanded_interaction_df` with additional
columns depending on the options specified in `metric_type`:
- `geo_min_dist_mean`: Mean of the minimum distances (km), included if
`"mean"` is selected.
- `geo_min_dist_median`: Median of the minimum distances (km), included if
`"median"` is selected.
- `geo_min_dist_mean_norm`, `geo_min_dist_median_norm`: Normalised versions,
included if `"norm"` (and `mean` or `median`, respectively) is selected.
- `geo_min_dist_mean_log`, `geo_min_dist_median_log`: Log-transformed
versions, included if `"log"` (and `mean` or `median`, respectively) is
selected.
}
\description{
This function computes distance-based metrics for each host-taxon-member –
hosted species interaction in `expanded_interaction_df`. For each species in
the host taxon, it calculates the minimum distance to all (other) hosts of
the same hosted species, optionally subsampling the coordinate data to
improve performance. Distances are precomputed in a lookup table to speed up
repeated queries, and the resulting metrics can be further transformed
(e.g. normalised or log-scaled) for downstream modelling.
}
\details{
- If a host or its associated hosts lack coordinate records, `NA` is returned
  for that interaction's distance metrics.
- Precomputation of distances for all host taxa improves performance for
  large datasets.
- Warnings are generated for taxa with missing geographic information or no
  other hosts.
}
\examples{
\dontrun{
# Generate example dataset
host_spp <- paste0("Quercus_", LETTERS[1:5])
hosted_spp <- paste0("Agrilus_", LETTERS[1:3])

interaction_df <- expand.grid(plant_sp = host_spp,
                              agrilus_sp = hosted_spp)

interaction_df$interaction <- c(1, 0, 0, 1, 0,
                                1, 0, 1, 0, 0,
                                0, 0, 0, 0, 1)

coords_df <- do.call(rbind, lapply(host_spp, function(sp) {
  n_ind <- sample(3:5, 1)
  data.frame(plant_sp = sp,
             lon = runif(n_ind, -10, 10),
             lat = runif(n_ind, 45, 55))
}))

hosts_df <- subset(interaction_df,
                   interaction == 1,
                   select = c("plant_sp", "agrilus_sp"))

# Calculate distance metrics for an example dataset
results <- generate_geo_metrics_df(expanded_interaction_df = interaction_df,
                                    coords_df = coords_df,
                                    known_interactions_df = hosts_df,
                                    host_taxon_col = "plant_sp",
                                    hosted_taxon_col = "agrilus_sp")
head(results)
}

}
