% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.r
\name{generate_pres_abs_df}
\alias{generate_pres_abs_df}
\title{Generate host presence/absence data frame for `caper::comparative.data()`}
\usage{
generate_pres_abs_df(observed_hosts, species_list)
}
\arguments{
\item{observed_hosts}{A character vector of observed host species.}

\item{species_list}{A character vector of all species of interest. Each
species in this list will appear in the output, with host.status 1 if
present in `observed_hosts` and 0 otherwise.}
}
\value{
A data.frame with two columns:
  \itemize{
    \item{species}{Species names from `species_list`.}
   \item{host.status}{Integer (0/1) indicating whether the species is a known host.}
  }
}
\description{
This function generates a presence/absence data frame for host species.
It takes a character vector of observed hosts and a full list of all
species, and returns a data frame marking species as hosts (1) or non-hosts
(0). Can be used to prepare data for input to `caper::comparative.data()`.
}
\examples{
\dontrun{
oak_hosts <- c("Quercus_robur", "Quercus_petraea")
all_oaks <- c("Quercus_rubra", "Quercus_robur",
              "Quercus_alba", "Quercus_petraea")
generate_pres_abs_df(oak_hosts, all_oaks)
#   species host.status
# 2 Quercus_rubra 0
# 3 Quercus_robur 1
# 4 Quercus_petraea 1
# 5 Quercus_alba 0
}

}
