% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{plot_phylogeny}
\alias{plot_phylogeny}
\title{Plot a phylogeny highlighting host species for a focal species}
\usage{
plot_phylogeny(
  phylo,
  host_spp = NULL,
  interaction_df = NULL,
  species = NULL,
  species_col = NULL,
  host_col = NULL,
  host_status_col = NULL,
  main = species,
  layout = "circular",
  offset = 5,
  size = 1.7,
  tip_colours = c("black", "coral"),
  legend_pos = "none"
)
}
\arguments{
\item{phylo}{An object of class \code{phylo} containing the phylogenetic
tree.}

\item{host_spp}{A character string specifying the host species. This can be
provided as an alternative to inferring the host species via `get_host_spp`
when an `interaction_df` is supplied.  Default: `NULL` (not required if
`interaction_df` is provided).}

\item{interaction_df}{A data frame containing species-host interaction data.
Default: `NULL` (not required if `host_spp` is provided).}

\item{species}{Character string giving the focal species of interest.}

\item{species_col}{A character string indicating the column name in
`interaction_df` that contains species names.  Default: `NULL` (not required
if `interaction_df` is provided).}

\item{host_col}{A character string indicating the column name in both `model`
and `interaction_df` that contains host species names.}

\item{host_status_col}{A character string specifying the column in
`interaction_df` that indicates the host status.  Default: `NULL` (not
required if `interaction_df` is provided).}

\item{main}{Character string indicating plot title. Defaults to the name
of the focal species (set to `NULL` to ignore).}

\item{layout}{Character string specifying the tree layout passed to
\code{ggtree} (e.g. "circular", "rectangular"). Default: `"circular"`.}

\item{offset}{Number indicating the offset of tip labels from the
phylogenetic tree. Default: `5`.}

\item{size}{Number indicating the size of tip labels. Default: `1.7`.}

\item{tip_colours}{Character vector of length 2, indicating colours for
non-hosts and hosts. Default: `c("black", "coral")`.}

\item{legend_pos}{Character string specifying legend position. Default:
`"none"`.}
}
\value{
A \code{ggplot} object produced by \code{ggtree}.
}
\description{
This function plots a phylogenetic tree and highlights host species
associated with a given focal species using tip label colours.
}
\examples{
plot_phylogeny(phylo = oak_phylo,
               interaction_df = interaction_data,
               species = "Agrilus coxalis",
               species_col = "agrilus_sp",
               host_status_col = "interaction",
               host_col = "quercus_sp")

}
\seealso{
\code{\link{get_host_spp}}
}
